+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## CHANGELOG
+### 0.5.0-prerelease.4
+Use [resize-observer-polyfill](https://github.com/que-etc/resize-observer-polyfill) to make sure the pager is always calculated with proper view dimensions
+
 ### 0.5.0-prerelease.3
 emit `viewChange` when `viewsToShow` options have changed